serverCode/requirements.txt
Flask==3.0.3
transformers==4.40.1
torch==2.3.0
scikit-learn==1.4.2
python-dotenv==1.0.1